+++ src/KOKKOS/fix_spring_rg_kokkos.cpp
@@ -135,9 +135,9 @@ void FixSpringRGKokkos<DeviceType>::operator()(TagFixSpringRGXcm, const int &i, 
 {
   if (mask[i] & groupbit) {
     Few<double,3> x_i;
-    x_i[0] = x(i,0);
-    x_i[1] = x(i,1);
-    x_i[2] = x(i,2);
+    x_i[0] = static_cast<double>(x(i,0));
+    x_i[1] = static_cast<double>(x(i,1));
+    x_i[2] = static_cast<double>(x(i,2));
     auto unwrap = DomainKokkos::unmap(prd,h,triclinic,x_i,image(i));
     const double massone = l_rmass_flag ? (double)rmass[i] : (double)mass[type[i]];
     result[0] += massone * unwrap[0];
